Realm.Sync.User -> Realm.User + put User in a separate file.
