+++ CHANGELOG.md
@@ -8,12 +8,13 @@ NOTE 1: Requires MacOS v10.14 (Mojave) to build.
 * Query-based Sync has been removed. This includes `Realm.Sync.Subscription`, `Realm.Results.subscribe()`, `Realm.subscriptions()`, `Realm.unsubscribe()`, `Realm.privileges()`, `Realm.Sync.User.isAdmin()`, `Realm.Sync.User.isAdminToken()`. These APIs are not initially supported by MongoDB Realm. They will be re-introduced in a future release. `partionKey` has been added to the `sync` configuration as a replacement. Read more [here](XXX). ([]())
 * Permissions has been removed. You need to configure permissions through Stitch.
 * Deprepated API has been removed: `Realm.openAsync()`, `Realm.Sync.setFeatureToken()`, `Realm.Sync.User.register()`, `Realm.Sync.User.adminUser()`, `Realm.Sync.User.registerWithProvider()`, `Realm.Sync.User.authenticate()`
+* `Realm.Sync.User` has been renamed to `Realm.User`.
 
 ### Enhancements
 * Added support for the `Decimal128` data type. ([#2674](https://github.com/realm/realm-js/issues/2674))
 * Added support for the `ObjectId` data type. ([#2675](https://github.com/realm/realm-js/issues/2675))
 * Added support for embedded objects. ([#2676](https://github.com/realm/realm-js/issues/2676))
-* Added `Realm.App`.
+* Added `Realm.App` and `Realm.Credentials`.
 
 ### Fixed
 * None.
