+++ binding.gyp
@@ -21,9 +21,11 @@
         }],
         ["realm_enable_sync", {
           "sources": [
+            "src/js_sync_util.hpp",
             "src/js_sync.hpp",
             "src/js_app.hpp",
             "src/js_app_credentials.hpp",
+            "src/js_user.hpp",
             "src/node/sync_logger.cpp",
             "src/node/sync_logger.hpp",
           ]
