+++ lib/extensions.js
@@ -349,11 +349,11 @@ module.exports = function(realmConstructor, context) {
     // Add sync methods
     if (realmConstructor.Sync) {
         let userMethods = require('./user-methods');
-        Object.defineProperties(realmConstructor.Sync.User, getOwnPropertyDescriptors(userMethods.static));
-        Object.defineProperties(realmConstructor.Sync.User.prototype, getOwnPropertyDescriptors(userMethods.instance));
-        Object.defineProperty(realmConstructor.Sync.User, '_realmConstructor', { value: realmConstructor });
-        realmConstructor.Sync.Credentials = {};
-        Object.defineProperties(realmConstructor.Sync.Credentials, getOwnPropertyDescriptors(userMethods.credentials));
+        Object.defineProperties(realmConstructor.User, getOwnPropertyDescriptors(userMethods.static));
+        Object.defineProperties(realmConstructor.User.prototype, getOwnPropertyDescriptors(userMethods.instance));
+        Object.defineProperty(realmConstructor.User, '_realmConstructor', { value: realmConstructor });
+        realmConstructor.Credentials = {};
+        Object.defineProperties(realmConstructor.Credentials, getOwnPropertyDescriptors(userMethods.credentials));
         realmConstructor.Sync.AuthError = require('./errors').AuthError;
 
         if (realmConstructor.Sync.removeAllListeners) {
@@ -377,7 +377,7 @@ module.exports = function(realmConstructor, context) {
             }
         }
 
-        setConstructorOnPrototype(realmConstructor.Sync.User);
+        setConstructorOnPrototype(realmConstructor.User);
         setConstructorOnPrototype(realmConstructor.Sync.Session);
 
         // A configuration for a default Realm
