+++ src/js_realm.hpp
@@ -439,6 +439,9 @@ inline typename T::Function RealmClass<T>::create_constructor(ContextType ctx) {
 
     FunctionType credentials_constructor = CredentialsClass<T>::create_constructor(ctx);
     Object::set_property(ctx, realm_constructor, "Credentials", credentials_constructor, attributes);
+
+    FunctionType user_constructor = UserClass<T>::create_constructor(ctx);
+    Object::set_property(ctx, realm_constructor, "User", user_constructor, attributes);
 #endif
 
     if (getenv("REALM_DISABLE_SYNC_TO_DISK")) {
